+++ internal/streams/in.go
