+++ internal/streams/out.go
