+++ internal/streams/stream.go
